tradingbot/requirements.txt
fastapi==0.68.2
uvicorn[standard]==0.15.0
requests==2.31.0
aiohttp==3.9.3
python-dotenv==1.0.0
# 数据库相关
motor==3.3.2
psycopg2-binary==2.9.9
redis==5.0.3
# 数据处理
numpy==1.26.4
pandas==2.0.3
scikit-learn==1.4.0
# 测试相关
pytest-cov==4.0.0
pytest-asyncio==0.23.5
nest-asyncio==1.6.0
pytest-mock==3.12.0
# 工具类
cachetools==5.3.2
beautifulsoup4==4.12.3
playwright==1.42.0
python-telegram-bot==21.0
prometheus-client==0.20.0
# NLP相关
spacy==3.7.4
nltk==3.8.1
torch==2.2.1
# 其他依赖
celery==5.3.6
email-validator==2.1.1
base58==2.1.1
praw==7.7.1
faker==24.8.0
python-jose[cryptography]==3.3.0
python-dateutil==2.9.0.post0
sqlalchemy==2.0.29
alembic==1.13.1
passlib==1.7.4
httpx==0.27.0
anyio==3.0.0
